backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
bcrypt==4.0.1
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.121.2
fastapi-cli==0.0.16
fastapi-cloud-cli==0.3.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
orjson==3.11.4
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyrefly==0.42.1
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
rsa==4.9.1
ruff==0.14.5
sentry-sdk==2.45.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uv==0.9.10
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
